+++ lib/View/booking/screen/office_filter_screen.dart
@@ -1,12 +1,11 @@
 import 'package:flutter/material.dart';
 import 'package:fluttertoast/fluttertoast.dart';
 import 'package:frontend/Presenter/booking/office_service.dart';
+import 'package:frontend/View/booking/component/custom_calendar.dart';
 import 'package:frontend/View/booking/component/select_time_button.dart';
-import 'package:frontend/View/colors.dart';
 import 'package:frontend/View/common/component/purple_bottom_button.dart';
 import 'package:frontend/View/common/component/sub_app_bar.dart';
 import 'package:intl/intl.dart';
-import 'package:table_calendar/table_calendar.dart';
 
 
 class OfficeFilterScreen extends StatefulWidget {
@@ -18,36 +17,12 @@ class OfficeFilterScreen extends StatefulWidget {
 
 class _OfficeFilterScreenState extends State<OfficeFilterScreen> {
 
-  static const weekdays = ['M', 'T', 'W', 'T', 'F', 'S', 'S'];
   static const startHintStr = '시작 시간을 선택해주세요';
   static const endHintStr = '종료 시간을 선택해주세요';
 
-  bool isShowTimePicker = false;
-
   DateTime? selectedDay;
-  DateTime focusedDay = DateTime.now();
-
   DateTime? startTime;
   DateTime? endTime;
-
-  @override
-  void initState() {
-    super.initState();
-
-    if (OfficeService().selectedDate != null) {
-      selectedDay = OfficeService().selectedDate;
-      focusedDay = selectedDay!;
-    }
-
-    if (OfficeService().startTime != null) {
-      startTime = OfficeService().startTime;
-    }
-
-    if (OfficeService().endTime != null) {
-      endTime = OfficeService().endTime;
-    }
-
-  }
 
   @override
   Widget build(BuildContext context) {
@@ -93,8 +68,6 @@ class _OfficeFilterScreenState extends State<OfficeFilterScreen> {
   }
 
   Widget renderCalendarView() {
-    TextStyle weekdayStyle = const TextStyle(fontSize: 11, color: purple);
-
     return Container(
       margin: const EdgeInsets.symmetric(vertical: 20),
       padding: const EdgeInsets.only(top: 10, left: 20, right: 20, bottom: 20),
@@ -102,41 +75,10 @@ class _OfficeFilterScreenState extends State<OfficeFilterScreen> {
         color: Color(0xFFF2F2F2),
         borderRadius: BorderRadius.all(Radius.circular(10))
       ),
-      child: TableCalendar(
-        rowHeight: 35,
-        focusedDay: focusedDay,
-        firstDay: DateTime(1800),
-        lastDay: DateTime(3000),
-        daysOfWeekHeight: 30,
-        headerStyle: const HeaderStyle(
-          formatButtonVisible: false,
-          titleCentered: true,
-          leftChevronVisible: true,
-          rightChevronVisible: true,
-          leftChevronMargin: EdgeInsets.only(left: 23),
-          rightChevronMargin: EdgeInsets.only(right: 23),
-          leftChevronIcon: Icon(Icons.chevron_left, color: purple,),
-          rightChevronIcon: Icon(Icons.chevron_right, color: purple,),
-        ),
-        calendarStyle: CalendarStyle(
-          isTodayHighlighted: true,
-          selectedDecoration: BoxDecoration(
-            color: purple.withOpacity(0.6),
-          ),
-        ),
-        calendarBuilders: CalendarBuilders(
-          dowBuilder: (context, day) {
-            return Center(child: Text(weekdays[day.weekday-1], style: weekdayStyle,),);
-          },
-        ),
-        onDaySelected: (DateTime selectedDay, DateTime focusedDay) {
-          setState(() {
-            this.selectedDay = selectedDay;
-            this.focusedDay = selectedDay;
-          });
-        },
-        selectedDayPredicate: (DateTime day) {
-          return isSameDay(selectedDay, day);
+      child: CustomCalender(
+        calendarDayHeight: 35,
+        changedDate: (DateTime date) {
+          setState(() { selectedDay = date; });
         },
       ),
     );
